requirements.txt
flask
openai
requests
python-dotenv
flask_cors
